Refactor for better background processing abstraction
- Move AppBase to app-base.ts for consistency
- Create a new lifecycle method on AppBase for background processing
- Return AppBase to top level controller instead of express object.
- Call background processing in the onListening callback
