+++ server/src/arcs-master-app.ts
@@ -6,7 +6,7 @@
 // subject to an additional IP rights grant found at
 // http://polymer.github.io/PATENTS.txt
 
-import {AppBase} from "./app";
+import {AppBase} from "./app-base";
 import {CloudManager} from "./deployment/cloud";
 import {Container, DeploymentStatus} from "./deployment/containers";
 import express from "express";
@@ -194,4 +194,4 @@ class ArcsMasterApp extends AppBase {
   }
 }
 
-export const app = new ArcsMasterApp().express;
+export const app = new ArcsMasterApp();
